% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_care.R
\name{combine_care}
\alias{combine_care}
\title{Combine care}
\usage{
combine_care(var1, var2)
}
\arguments{
\item{var1}{variable at admission}

\item{var2}{variable at discharge}
}
\value{

}
\description{
Specific function that combine 'twin' variables of patient's care at admission (\code{var1}) and at discharge (\code{var2}) in to a single variable. Input variable is coded with values 'Yes' or 'No'. Output variable is code with values 'Yes', 'No at admission then not reported', 'No at any time' or 'Not reported'.
}
