% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{country_plot}
\alias{country_plot}
\title{Plot cases and deaths and trends into a single graphic plot}
\usage{
country_plot(
  dta,
  df_trends,
  cases_lab = "cases",
  deaths_lab = "deaths",
  ylab_curve = "Frequency",
  title_curve = "Since the first cases reported",
  title_mdl1 = "Last 30 days",
  title_mdl2 = "Last 14 days",
  ylab_mdl = "Frequency and fitted values",
  caption = "Fitted using a linear regression model",
  add_title = FALSE,
  title = "COVID-19 case and death trend estimations in"
)
}
\arguments{
\item{dta}{daily case and death data for country}

\item{df_trends}{trend preds etc for country}

\item{add_title}{add a title to plot?}
}
\value{
ggplot
}
\description{
Plot cases and deaths and trends into a single graphic plot
}
