% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{country_table}
\alias{country_table}
\title{gt country summary table}
\usage{
country_table(df_trends)
}
\arguments{
\item{df_trends}{filtered to one country}
}
\value{
gt table
}
\description{
gt country summary table
}
