% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_owid_jhcsse.R
\name{get_country_summaries}
\alias{get_country_summaries}
\title{Get summary df of country cases, deaths and trends etc}
\usage{
get_country_summaries(df_jhu, inc_factor = 1e+05)
}
\arguments{
\item{df_jhu}{dataframe output from \code{\link[=get_owid_jhcsse]{get_owid_jhcsse()}}}

\item{inc_factor}{incidence factor. defaults to 100 000}
}
\value{
tibble
}
\description{
Get summary df of country cases, deaths and trends etc
}
