% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecdc.R
\name{get_ecdc_data}
\alias{get_ecdc_data}
\title{Import ECDC data}
\usage{
get_ecdc_data()
}
\value{
a tibble dataframe
}
\description{
https://opendata.ecdc.europa.eu/covid19/casedistribution/csv
}
\details{
Performs standardisation of country names and infers ISO3, continent and region
data using the \code{\link[countrycode:countrycode]{countrycode::countrycode}} function
}
\examples{
df_ecdc <- get_ecdc_data()
}
