% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msf_linelist.R
\name{get_msf_linelist}
\alias{get_msf_linelist}
\title{Import MSF data from a remote location}
\usage{
get_msf_linelist(
  path_local = path.local.msf.data,
  file_name = "dta_MSF_linelist.RDS",
  path_remote = path.sharepoint.data,
  force = FALSE
)
}
\arguments{
\item{force}{}
}
\value{

}
\description{
Import MSF data from a remote location
}
