% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_owid_jhcsse.R
\name{get_owid_jhcsse}
\alias{get_owid_jhcsse}
\title{Import JHU CSSE data from Our World in Data's github}
\usage{
get_owid_jhcsse(
 
    url = "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/jhu/full_data.csv"
)
}
\arguments{
\item{url}{raw OWID github URL for JHU CSV}
}
\value{
a tibble dataframe
}
\description{
Download URL: https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/jhu/full_data.csv
}
