% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_owid_jhcsse.R
\name{get_owid_summary}
\alias{get_owid_summary}
\title{Import OWID covid summaries per country}
\usage{
get_owid_summary(
 
    url = "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/latest/owid-covid-latest.csv"
)
}
\arguments{
\item{url}{OWID covid latest github URL}
}
\value{
a tibble dataframe
}
\description{
Download URL: https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/latest/owid-covid-latest.csv
}
