% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_trends}
\alias{get_trends}
\title{Compute case and deaths trends}
\usage{
get_trends(
  df,
  time_unit_extent = 14,
  omit_past_days = 2,
  add_ci = TRUE,
  add_dt = TRUE,
  add_preds = TRUE
)
}
\arguments{
\item{df}{JHSU CSSE data}

\item{time_unit_extent}{defaults to 14 days}
}
\value{
tibble
}
\description{
Compute case and deaths trends
}
