% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_breaks.R
\name{label_breaks}
\alias{label_breaks}
\title{Label breaks}
\usage{
label_breaks(breaks, exclusive = FALSE, add_Inf = TRUE, replace_Inf = TRUE)
}
\arguments{
\item{breaks}{a vector with break values (\code{Inf} may be included)}

\item{exclusive}{logical, indicating whether the right side of the label should display a mutually exclusive value with the left side of the following label. If FALSE (default), the labels display not mutually exclusive values (i.e. "1-10", 10-100", etc.). If TRUE the labels display mutually exclusive values (i.e. "1-9", 10-99", etc.).}

\item{add_Inf}{logical, if TRUE (default), and if not already present, the value \code{Inf} is added as last value. If TRUE the length of the output vector is of the same input vector. If FALSE the length of the output vector is the length of the input vector -1.}

\item{replace_Inf}{logical, if TRUE (default) the value \code{Inf} is replaced by a +.}
}
\value{
a vector of the same length of the input vector of length of the input vector - 1 if \code{add_Inf} is set as FALSE.
}
\description{
Create two-sided labels from a sequence of values
}
\examples{
x <- c(1, 10, 100, Inf)
label_breaks(x)
label_breaks(x, exclusive = TRUE)
label_breaks(x, exclusive = FALSE, replace_Inf = TRUE)

y <- c(1, 10, 100)
label_breaks(y)
label_breaks(y, exclusive = TRUE) # same length of input vector
label_breaks(y, exclusive = FALSE, add_Inf = FALSE) # length of input vector - 1
}
