% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday_week_date.R
\name{monday_week_date}
\alias{monday_week_date}
\title{Return a date variable set to the Monday of the week}
\usage{
monday_week_date(date)
}
\arguments{
\item{date}{a date}
}
\value{

}
\description{
Function that return for any day of the week a date set to the Monday of that week.
}
