% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ecdc_data.R
\name{prepare_ecdc_data}
\alias{prepare_ecdc_data}
\title{Prepare ECDC data}
\usage{
prepare_ecdc_data(dta)
}
\value{
a tibble dataframe
}
\description{
Performs standardisation of country names and infers ISO3, continent and region data using the countrycode::countrycode function
}
\examples{
df_ecdc <- get_ecdc_data()
}
