% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_msf_data.R
\name{prepare_msf_dta}
\alias{prepare_msf_dta}
\title{Prepare MSF Covid linelist data}
\usage{
prepare_msf_dta(dta, shorten_var_names = FALSE)
}
\arguments{
\item{dta}{The raw linelist downloaded}

\item{shorten_var_names}{Shorten the names of the variables by removing suffixes such as "MSF_", "patinfo_", "patcourse_"}
}
\value{

}
\description{
This function clean incoherences, standardise/factorise main variables and create new variables that are used in different analyses
}
