% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecdc_data.R
\name{get_ecdc_data}
\alias{get_ecdc_data}
\title{Import ECDC data}
\usage{
get_ecdc_data()
}
\value{
a tibble dataframe
}
\description{
get data from https://opendata.ecdc.europa.eu/covid19/casedistribution/csv or from https://www.ecdc.europa.eu/sites/default/files/documents/COVID-19-geographic-disbtribution-worldwide.xlsx
}
